Add paragraph tag to link example
This link example does not include a wrapping element, meaning that some of the text is appearing as default serif text instead of in Transport. As the text above this example states that it is
regarding links at the end of sentences or paragraphs, the example shouldn't be negatively affected by the inclusion of a paragraph.