Allow start buttons to be forms
Although unusual, there are a few valid circumstances where you might
want to submit a form when pressing the start button. A situation that
came up recently was we wanted to record analytics on when a user
started a service by recording the time that a user presses the start
button, without client-side JavaScript. The implementation for this
involved using the Rails `button_to` helper which generates a form which
POSTs to a route and includes a single `submit` button.

I think this is a valid use case of the start button, so I suggest
changing the wording on this to be slightly more open to the idea of
the start button not being a link.