Remove examples for archived ethnic groups pattern
The ‘Ask users for ethnic groups’ pattern was archived in #1582 [1] (dd77104) when the guidance was moved into the ‘Ask users for equality information’ pattern.

New examples were created for the equality information pattern, but the existing examples and the SVG included in the guidance were not removed from the repo.

Remove the examples and the SVG previously included in the archived ethnic groups pattern.

[1]: https://github.com/alphagov/govuk-design-system/pull/1582