Fix matchMedia event listener in IE / Safari < 14
IE 10/11 and Safari >= 5.1 < 14 support `window.matchMedia` but do not support `MediaQueryList.addEventListener`.

Instead we have to use the non-standard (and deprecated) `MediaQueryList.addListener`, which only takes one argument which is the callback function to run when the media query status changes [1].

Test for the presence of the `MediaQueryList.addEventListener` method and then use whichever method is appropriate for the browser.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/MediaQueryList/addListener

Co-authored-by: Owen Jones <64783893+owenatgov@users.noreply.github.com>