Retry failed tests
The "see no evil, hear no evil" approach. This does not address the flaky tests themselves, but sets `jest.retryTimes` to 3, rerunning failed tests and greatly increasing the likelihood of a CI test task passing.

This is speedier (and more automatic) than rerunning a full CI task again, so might be enough if the flakiness isn't easily addressable.