Update index.md.njk
New static classes in 4.3.0 css assets have the phrase 'static' towards the end of the class name rather than the beginning. eg.

.govuk-\!-margin-static-1 {
  margin: 5px !important
}

The sentence above the examples 'The static spacing override classes start with `govuk-!-static`. Use them the same way as the responsive spacing override classes.' would also need to be changed as it states the class starts with static. That is unless the guidance is actually correct and it's the generated assets that are back to front?