Stop recommending pattern=[0-9]* on number inputs
We previously recommended using `pattern="[0-9]*"` on number inputs to prompt iOS to display the numeric keypad.

This has been unnecessary since Safari 12.2, when support for the standardised `inputmode` attribute was added to Safari.

We stopped using the `pattern` attribute on the date input component [1] in v4.1.0 (released in May 2022) as the proportion of GOV.UK visitors using versions of iOS 12.x and below had fallen to 0.08% of total traffic.

[1]: https://github.com/alphagov/govuk-frontend/pull/2599

Co-authored-by: Calvin Lau <77630796+calvin-lau-sig7@users.noreply.github.com>