Add border to 'preview' / 'archive' banner
We moved the banner when refactoring the nav in 2b6a27b.

Now that the banner is not between the header and the navigation, it makes sense to add the bottom border back to separate the banner from the page content.

Keep the `--no-border` modifier which we use on the homepage when the banner is followed immediately by the masthead.