Remove redundant app-pane wrapper
When the Design System had an independently scrollable nav and body regions, we used a 'app-pane' flexbox wrapper around the whole site so that the header could be fixed in place above both scrollable regions.

When we simplified the layout and lost the indepedantly scrollable regions we didn't tidy up the wrapper, which isn't really doing anything any more (other than the flex layout causing weird issues like the banner width issue seen in 6c83fde).

Remove the unneccesary wrapper and its classes.