Remove custom phase banner styles
The `width: 100%` is no longer needed that we're not in a flex context and can be removed.

As it turns out, the margins are being overridden by the `app-width-container` class anyway – and if they weren't being overriden, they'd mean that the phase banner would be flush up against the viewport edges on mobile, which isn't what we want.