Remove use of govuk-exports
We need to use govuk-exports in GOV.UK Frontend as files can be imported multiple files but we only ever want the CSS to be outputted once.

As we're in control of how and where Sass in the Design System is imported, we don't need to worry about this and don't need to use the exports function.