Bump jstransformer-marked from 1.0.3 to 1.2.0
The calm before the storm...

We use `metalsmith-in-place` to compile our markdown and nunjucks. That in turn has a dependency on `jstransformer`, which then requires peerDependencies of the relevant packages - in this case, `jstransformer-marked` and `jstransformer-nunjucks`.

So despite us having `marked` 4.0.18 as a dependency, we're at the whim of `jstransformer-marked`'s dependency tree when it comes to processing our files with `metalsmith-in-place`.

v1.0.3 of `jstransformer-marked` uses `marked` ^0.3.9 (0.3.19 in our package-lock).

v1.2.0 uses `marked` 4.0.18.

Surely this won't be an issue... (spoiler: it will).