Remove idPrefix from examples
I've found an example in the GOV.UK Account contact page where the idPrefix and name went out of sync leading to error summary not linking up nicely to the first input.
I think to lower the likely hood of this in the future we could make the examples use the `name` implicitly by default.
What do you think? I can update the other examples if you like this proposal.