Remove references to pattern attribute in guidance
We stopped recommending using the `pattern` attribute on number inputs in #2323 but I missed this reference in the guidance for the text input component and the bank details pattern.

We still tell users to turn off HTML validation in the broader 'validation' pattern – that hasn't changed – but now that we're not using the pattern attribute there's no need to call it out in these specific places.