Remove `idPrefix` from examples
`idPrefix` is not required as `name` option will be used by default.

> String to prefix id for each checkbox item if no id is specified on each item.
> If not passed, fall back to using the name option instead.

This change is based on findings that when users change their `idPrefix`
they can often forget to change their error handling to account for this
when outputing links in their Error Summary component.

By keeping these in sync by default it lowers the risk of breaking that relationship.