Update jstransformer dependencies
These are used by `@metalsmith/in-place` to transform Nunjucks and Markdown.

Bumping `jstransformer-nunjucks` results in no diff to the build output.

Bumping `jstransformer-marked` updates the version of Marked being used, which in turn massively mangles our output.

Limiting to patch bumps for now until we can sort the markdown output.