Bump jstransformer-marked to 1.2.0
We use `metalsmith-in-place` to compile our markdown and nunjucks. That in turn has a dependency on `jstransformer`, which then requires transformers of the required type (in this case, `jstransformer-marked`).

We have `marked` as a direct dependency, but we're actually at the whim of `jstransformer-marked`'s dependency tree when it comes to processing our files with `metalsmith-in-place`.

v1.0.3 of `jstransformer-marked` uses `marked` ^0.3.9 (0.3.19 in our package-lock).

v1.2.0 uses `marked` 4.0.18.

This creates... problems.