Add autolinks to markdown.
Pedantic rules require angle brackets around raw URLs in order to autolink them. This was a bit buggy in early versions of Marked, which is why it's flown under our radar till we bumped Marked.

Switching to Github Flavoured Markdown should negate the need to do this and allow future editors not to think about it, but it's probably still worth being widely compliant in case anything changes down the line.