Update notification banner to disable autofocus in JS
We don't want the examples for the notification banner in the Design System to take focus as it's not useful for users of assistive tech in this context, and it causes the viewport to scroll to the first example on page load.

This means we need to disable the auto focus feature, but we don't want it to be disabled in the code blocks that we provide under the examples – otherwise when teams copy and paste this code they'll end up disabling auto focus where it's used in services.

Previously we achieved this by having two versions of the code:

- `index.njk` which is used to generate the HTML used for the interactive example of the component, and includes the `disableAutoFocus` option
- `code.njk` which is used to generate the Nunjucks and the HTML displayed in the code blocks under the example, which does not include the the `disableAutoFocus` option (because we don't want users to copy it)

Now that GOV.UK Frontend v4.4.0 is out we have a simpler way to achieve the same thing, by changing the config in the JS [1]. This means that we can simplify each example by deleting `index.njk` and then renaming `code.njk` to` index.njk`.

This reduces the risk of the two versions of the code will get out of sync.

As it happens, we need to fix the fact that they've _already_ got out of sync when the `<p>` was changed to a `<h3>` in 858a3e7 (#1515) but the same change was not made to `index.njk`.

[1]: https://github.com/alphagov/govuk-frontend/blob/main/CHANGELOG.md#configure-components-in-javascript