Use correct variable name for colours from Sass
We've been working around a bug in sass-export that has now been fixed [1], where a variable prefixed with an underscore gets the underscore replaced with a dash.

So where we were looking for `$-govuk-colour-palette-modern` instead of the correct variable name`$_govuk-colour-palette-modern`, we now need to switch back to use an underscore to get the tests passing again.

[1]: https://github.com/plentycode/sass-export/issues/77