[NOTE] eleventy successfully compiles
eleventy now successfully compiles to `deploy/public`, with a bunch of quirks.

Next steps are to squash the diffs between what the current compiled website looks like, and what eleventy produces.

Following that, refactoring the whole lot and getting rid of unneeded code [might be beyond the scope of this spike]