Compress SVG and ICO files with gzip
Running the Design System homepage through WebPageTest [1] flagged that we’re not currently compressing SVG or ICO files:

> FAILED - (17.7 KB, compressed = 2.3 KB - savings of 15.4 KB) - https://design-system.service.gov.uk/images/homepage-illustration.svg
> FAILED - (6.2 KB, compressed = 2.5 KB - savings of 3.7 KB) - https://design-system.service.gov.uk/assets/images/favicon.ico

We also have a number of SVGs on other pages – including the community page – where we’ll see a benefit:

```
/patterns/check-a-service-is-suitable/check-a-service-is-suitable-new.svg
/images/homepage-illustration.svg
/community/images/dsday22-logo.svg
/community/images/discuss-and-give-feedback.svg
/community/images/propose-a-change-to-pages.svg
/community/images/share-your-examples-and-research.svg
/community/images/chat-with-us-on-slack.svg
/assets/images/govuk-mask-icon.svg
```

Other best practise guides for configuring gzip for nginx include both `image/svg+xml and `image/x-icon` in the list of mime types to gzip.

Add them to our list of mime types so users can benefit from the extra compression. Sort the list and multi-line it so that it’s easier to read.

[1]: https://www.webpagetest.org/result/230106_BiDc3X_9B5/3/performance_optimization/