Exclude headings from excluded sections
When adding documents to the search index, we do a separate pass of the files to add headings from any documents with `show_page_nav: true` as separate entries, so that searching for things like ‘Links’ (a section within the ‘Typography’ page) works.

However, unlike when adding documents, this pass did not exclude documents that were in sections excluded from search in `config/navigation.json`.

The new accessibility strategy page in the community section uses `show_page_nav: true` which means that the H2s within the accessibility strategy are currently showing up in search results. This is inconsistent with the rest of the community section which is currently excluded from search.

Refactor the code to do the ‘common’ filtering of files first before creating separate sets of document and page heading results to index, ensuring that both passes take into account the section the documents are in.