Add example for showing a field that was optional
The guidance for check your answers mentions hiding sections that the user doesn't see but not optional questions. On discussion in cross-gov slack, teams do play back said fields but say "not provided". This may need to be added to guidance - view separate note - but also would be helpful to change the example