Change GOV.UK Frontend to 'fully' compliant
I had meant to update this form 'mostly' to 'fully' a while back, but forgot.

The GOV.UK Frontend doesn't have any known WCAG 2.1 AA failures, so it is considered fully compliant. We do list one known and confirmed accessibility issue, but the issue is not a WCAG 2.1 AA failure.