Fixed the section on 'not within the scope'
The model accessibility statement requires us to list pieces of our products which are "not within the scope of the accessibility regulations". We don't have any that we claim, so I've added an explicit note on that.

Also changed the heading of the next section, so it doesn't get confused with this section. 
Before: Accessibility concerns not within the scope of the accessibility regulations
After: Other identified and tracked accessibility concerns