Bump typography changes from v5 on roadmap
We’ve decided to roll the typography changes out behind a feature flag as part of a future 5.x release, rather than trying to include them in v5.

We want to give ourselves a little more time to properly understand the impact of these changes.

We also think that putting the changes behind a feature flag will make them easier to roll out, before we then enable them by default, likely in v6.