Include a section on chart colours
I have added in this change, as the problem with using the GOV.UK design system colours which are already on this page for charts is that, depending on what order the colours are presented in, they would not meet the adjacent colour contrast requirements.

The colours guidance I have linked was developed by a cross government working group led by the Department for Education. The colour palette in this guidance have been designed for use in standard line, bar and pie charts. When all the guidance is followed, the colours mentioned in the Analysis Function colour palette will create charts which are as accessible as possible. This  colour palette is used in the Whitehall publisher on GOV.UK.