Turn legacyMode off
`legacyMode` uses iframes to communicate between Axe and the page
https://github.com/dequelabs/axe-core-npm/tree/develop/packages/puppeteer#axepuppeteersetlegacymodelegacymode-boolean--true

This causes issues with unloaded iframes. When we ran `iframe-resizer` on each Example as it initalised, everything was hunky dory. Now that we're only running `iframe-resizer` on each lazily loaded iframe, axe gets confused.

The solution is to simply turn off `legacyMode`.

There are some concerns that doing so makes things slow when running lots of axe tests, but it doesn't look like our small test suite is very affected.