Fix missing Expires headers for JavaScript
In b16393df117a0b499b6288ea0041054093f94d58 We changed the MIME type for JavaScript from `application/javascript` to `text/javascript` in line with IETF RFC 9239 [1].

However we missed that the caching logic needed updating too, so we’ve not been sending `Expires` headers for our JS since. This means that the JS is not currently being cached by CloudFront or in browsers.

Update the mime type to match the mime type now being served.

[1]: https://www.rfc-editor.org/rfc/rfc9239