Stop deploying to PaaS
We’ve migrated the Design System from PaaS (which is being decommissioned [1]) to Netlify.

We no longer need to deploy the Design System to PaaS, so we can remove the deploy workflow (as well as the ability to upload the built site as an artefact from the test workflow which meant we could avoid rebuilding it as part of the deploy job).

Closes #3342.

[1]: https://gds.blog.gov.uk/2022/07/12/why-weve-decided-to-decommission-gov-uk-paas-platform-as-a-service/