re-route archived link on "Getting started: production" page
It looks like the old link here was archived in November / December, as part of the 5.0 release. I feel like routing people to the Frontend docs page on making sure to keep up to date with the latest version is a pretty valuable alternative piece of guidance for teams still stuck on GOV.UK Elements, Templates, etc.