Work around module.exports not being type checked
There’s an issue with the TypeScript compiler that prevents direct assignments to `module.exports` being type checked. [1]

We can work around this by assigning the config to a const first, then exporting it as a second step.

This means that the invalid config is correctly flagged when running `npm run lint:types` and in VS Code.

[1]: https://github.com/microsoft/TypeScript/issues/47107