Remove non-existent option from nav config
The logic for including pages in search was flipped in #3110 (76dcb10) whilst the work on the new accessibility section was happening in a separate branch.

As she default is now for pages to be included in search unless `ignoreInSearch` is true, so this config no longer does anything.

This was missed when rebasing and merging the branch in #3090.

Remove the redundant and non-existent config option to avoid confusion.