Draft example API of how we could initialise components
First things first, this code has not even been run, it's only a draft of an API I think could be nice for initialising components, accounting for the nuances found in the different components of the Design System.

- Some components run on a specific element, some not. So the API need to allow both.
- When initialised on a specific element, the module name could be a static property of the component, however, it's probably useful to let it be set explicitely
- Further selection, like finding the  for the Copy, should probably be part of the component's responsibility, rather than the initialisation code
- Similarly, if a component needs to be initialised only once, this could be part of the component's responsibility, rather thatn the initialisation loop.

TypeScript might get in the way of having something so dynamic. Especially, I'm not 100% sure how we can note that a function accepts a 'class' as parameter. Seems it's a bit tricky based on this: https://github.com/Microsoft/TypeScript/issues/17572