Add documentation for 'id' to accordion component
The extra documentation for the 'id' attribute that was present in the
Nunjucks options table for the accordion has been added to the 'How It
Works Section'. This is because if a user is just using the HTML of the
accordion component, then they might not encounter the additional
documentation provided for the 'id' attribute.

This addresses issue #1073.