Improve link text in phase banner examples
The link text in our examples isn’t very good. It’s just the word ‘feedback’ that’s linked, which is ambiguous and so it's unclear where the page takes you – it could be a way to view feedback that the service has received, for example.

As the surrounding paragraph does provide enough context to make it clear I believe this still meets WCAG 2.2 2.4.4: Link Purpose (In Context) (A) but would not be sufficient to pass 2.4.9 Link Purpose (Link Only) (AAA).

Improve the link text by making more of the sentence part of the link, and also make it clear how the user will give feedback (by email, or in a form that opens in a new tab).

Restructure the sentence to put the link at the end of the sentence.