Add Embed Card component
Update `application.mjs` to support multiple cookie banners on a page.

Create new Embed Card component. If require cookie not set (or script
fails to execute) placeholder image that links to YouTube video will be
created. If cookie is set then placeholder will be replaced with a YT
embed.

Add `lazyEmbedObserver` in `application.mjs`. Embed Card component will
only be initialized when the placeholder is in the viewport. This means
for a page with lots of embeds, the page the performance hit will be
limited if the user doesn't scroll the entire page.

Add tests for new Embed Card component.