Create typeText function to control keyboard typing
Flakiness in the search tests is caused by some kind of timing/race condition with the .type() method not completing before we assert our expectations.

Adding a delay to the text input seems to make it less flaky, but I've moved into a separate function to make it easier to iterate fixes if further flakiness occurs.

I've tried sending characters individually, setting focus on each character, and using page.waitForFunction, but this seems to be the least flaky fix.