Split out excludes, and globs
It might be an axe bug, but exclude doesn't seem to work properly with an array argument. Splitting them out reduces failures from 300+ to 139.

Additionally, I've made some changes to which files are included in the glob, with a view to splitting these out and only running certain tests in certain environments to avoid jammingthings up for too long.