Update index.md
I work in the Copyright Team at The National Archives, which is responsibility for Crown copyright and its licensing.  I have amended the text of the "How it works" section to add a bit more information for content managers as we recently had an enquiry about the copyright text in the footer.