Use same `order` property for sorting top level nav and sub nav
This makes all navigation items be sorted on the same property for consistency.

To identify initial sections, we can rely on the `showSubNav` property,
though we could use a custom `includeInNavigation` property which would trigger
both the page being listed in the top-level navigation and the rendering of the sub navigation.