Move control of overall padding of side pane content
The fact that we needed to apply padding to both `app-breadcrumbs`
and `app-content` was a sign that the padding was not applied
to the right elements and should instead be applied to their parent.

This means that whatever gets added inside `app-side-pane__content`,
the right amount of space will surround it.