Further normalise how vertical padding of site's contents is applied
Moves the vertical padding in the white area where the content is displayed
to a `app-main-wrapper` class. This ensures the spacing is consistent
between `layout-pane` and `layout-single-page` so the breadcrumbs end up
at the same level on all pages.

The padding is applied as its own class in a similar fashion that we have
`govuk-main-wrapper` in GOV.UK Frontend, but is made to match the spacing
designed for the `layout-pane`.

It also fixes the bottom padding having disappeared after the content
on the pages using `layout-single-page`.