Bump npm version in package.json
We’re currently having issues with Dependabot updates timing out.

GitHub support have advised us that specifying a version of npm >= 10.5.0 might help with the timeouts:

> Fortunately, I've received an update from the Dependabot Updates Software Engineering Team today.
>
> They have pointed out that this issue appears to be related to the corepack npm install handling of requests to registry.npmjs.org.
>
> This is a known npm issue that affects npm version 10 up to 10.5.0.
>
> In your package.json you are specifying npm version 10.1.0 or greater. They've suggest explicitly specifying version 10.5.0 or greater instead.

We can also see some indications that this might be the problem in the Dependabot logs from a recent run.

Initially we see that npm@10.5.0 is being used:

https://github.com/alphagov/govuk-design-system/actions/runs/16417136291/job/46385737171#step:3:135

The engine constraints are then processed – which initially results in an error about it being an illformed requirement – before it successfully determines the constraints and installs npm@10.1.0:

https://github.com/alphagov/govuk-design-system/actions/runs/16417136291/job/46385737171#step:3:150

This completes successfully, which suggests that npm@10.5.0 can successfully talk to the registry.

This process is repeated several times for reasons I do not understand.

After that point Dependabot does seem to be able to get some information from the registry, but many steps time out.

Try following the advice from GitHub support and bump the version of npm to 10.5.0 in our engines config.