Downgrade iframe-resizer to 4.3.11
We currently see this message in the console on the Design System site:

```
IFRAME-RESIZER

Iframe-Resizer 5 is now available via the following two packages:

 * @iframe-resizer/parent
 * @iframe-resizer/child

Additionally their are also new versions of iframe-resizer for React, Vue, and jQuery.

Version 5 of iframe-resizer has been extensively rewritten to use modern browser APIs, which has enabled significantly better performance and greater accuracy in the detection of content resizing events.

Please see https://iframe-resizer.com/upgrade for more details.
```

We are not planning to upgrade to v5 as the license model has changed to GPL v3, which means using it would require us to license our own code under GPL v3 too (or pay for a commercial license).

There is no way to disable this message from being logged by the dependency.

Looking at the diff between 4.3.11 and 4.4.5 [1], there are no meaningful changes to iframe-resizer, only README changes and the introduction of these upgrade prompts.

Downgrade to 4.3.11 to remove the prompts, and add a test that should hopefully stop people from upgrading it again.

[1]: https://github.com/davidjbradshaw/iframe-resizer/compare/v4.3.11...v4.4.5