Ignore all iframe-resizer updates
We recently downgraded iframe-resizer from 4.4.5 to 4.3.11 in 85cc5fc.

There are no meaningful changes to iframe-resizer between 4.3.11 and 4.4.5 – only README changes and the introduction of some annoying upgrade prompts in the console.

We are also not planning to upgrade to v5 as the license model has changed to GPL v3, which means using it would require us to license our own code under GPL v3 too (or pay for a commercial license).

Tell Dependabot to ignore all updates to iframe-resizer, not just the major version (bump to v5).