Fix VoiceOver listing mobile menu toggles on desktop
Adds an `aria-hidden` attribute in complement to the `hidden`
attribute already added to the toggles on desktop viewports.

While in there, also adds a `disabled` attribute while the toggles
are hidden, to prevent their activation in case another bug keeps
them accessible (and to prevent their activation through `.click()`
in JavaScript).