Abstract common browsersync options
We have two ‘tasks’ that use browsersync, ‘serve’ and ‘start’.

Both are currently setting the same config options.

Move the common options into a separate config file.

Although port is only currently defined in the ‘serve’ task, define that in the common config as well as the default configured port is the same as browsersync default port. This aligns the behaviour between the two tasks, so that e.g. the `PORT` environment variable can be used in both places to override the port.