Remove check-external-links script
The check-links script now checks all URLs except GitHub, which mimics the behaviour of the check-external-links script, and does it in seconds.

It's possible to add retries and include Github urls, but it's a bit flaky because of rate limiting. I at least ran it and discovered only 429 status codes.

But this is a gap in our testing that would be nice to plug.